bloom: add adaptive filter policy
Add an adaptive bloom filter which is configured with a max size. The
bits per key are reduced as necessary so the filter fits within the
given size. The filter format does not change and remains readable by
older versions.

We plan to use this to generate smaller filters in L6 (in a future
change).

+++ data_test.go
@@ -1697,7 +1697,7 @@ func parseDBOptionsArgs(opts *Options, args []datadriven.CmdArg) error {
 			if err != nil {
 				return err
 			}
-			fp := bloom.FilterPolicy(v)
+			fp := bloom.FilterPolicy(uint32(v))
 			opts.TableFilterDecoders = []TableFilterDecoder{bloom.Decoder}
 			for i := range opts.Levels {
 				opts.Levels[i].TableFilterPolicy = fp
